Update for version 1.0

+++ indexer_JPIP/cio.c
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2001-2002, David Janssens
  * Copyright (c) 2003, Yannick Verschueren
- * Copyright (c) 2003,  Communications and remote sensing Laboratory, Universite catholique de Louvain, Belgium
+ * Copyright (c) 2003, Communications and remote sensing Laboratory, Universite catholique de Louvain, Belgium
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -109,12 +109,14 @@ void cio_write(long long v, int n) {
 /// <summary>
 /// Read some bytes.
 /// </summary>
-unsigned int cio_read(int n) {
+/* unsigned int cio_read(int n) { */
+long long cio_read(int n) {
     int i;
-    unsigned int v;
+    /*unsigned int v;*/
+    long long v;
     v=0;
     for (i=n-1; i>=0; i--) {
-        v+=cio_bytein()<<(i<<3);
+      v+=cio_bytein()<<(i<<3);
     }
     return v;
 }
